docs(*.py): Update types

+++ circle.py
@@ -6,10 +6,10 @@ def area(r):
     Return area of circle.
 
         Parameters:
-            r (double): Radius of circle.
+            r (float): Radius of circle.
         
         Return value:
-            area (double): Area of circle.
+            area (float): Area of circle.
     '''
     return math.pi * r * r
 
@@ -19,10 +19,10 @@ def perimeter(r):
     Return perimeter of circle.
 
         Parameters:
-            r (double): Radius of circle.
+            r (float): Radius of circle.
         
         Return value:
-            perimeter (double): Perimeter of circle.
+            perimeter (float): Perimeter of circle.
     '''
     return 2 * math.pi * r
