+++ rectangle.py
@@ -3,11 +3,11 @@ def area(a, b):
     Return area of rectangle.
 
         Parameters:
-            a (double): Length of first side of rectangle.
-            b (double): Length of second side of rectangle.
+            a (float): Length of first side of rectangle.
+            b (float): Length of second side of rectangle.
         
         Return value:
-            area (double): Area of rectangle.
+            area (float): Area of rectangle.
     '''
   return a * b
 
@@ -16,10 +16,10 @@ def perimeter(a, b):
     Return perimeter of rectangle.
 
         Parameters:
-            a (double): Length of first side of rectangle.
-            b (double): Length of second side of rectangle.
+            a (float): Length of first side of rectangle.
+            b (float): Length of second side of rectangle.
         
         Return value:
-            perimeter (double): Perimeter of rectangle.
+            perimeter (float): Perimeter of rectangle.
     '''
   return (a + b) * 2
